requirements.txt
asgiref==3.9.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
crispy-bootstrap5==2025.6
cryptography==46.0.1
defusedxml==0.7.1
dj-database-url==3.0.1
Django==4.2.24
django-allauth==0.57.2
django-crispy-forms==2.4
django-widget-tweaks==1.5.0
gunicorn==20.1.0
idna==3.10
oauthlib==3.3.1
psycopg2-binary==2.9.10
pycparser==2.23
PyJWT==2.10.1
python3-openid==3.2.0
requests==2.32.5
requests-oauthlib==2.0.0
setuptools==80.9.0
sqlparse==0.5.3
urllib3==2.5.0
whitenoise==6.11.0
